+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 0.20.0
+- Refactor game.dart classes into separate files
+- Split up gameloop into separate GameLoop class
+- Adding sprites example
+- Made BaseGame non-abstract and removed SimpleGame
+- Adding SpriteButton Widget
+- Added SpriteBatch API, which renders sprites effectively using Canvas.drawAtlas
+- Introducing basic effects API, including MoveEffect and ScaleEffect
+
 ## 0.19.1
  - Bump AudioPlayers version to allow for web support
  - Adding Game#pauseEngine and Game#resumeEngine methods